Move `GameDriver` to dedicated module
