+++ src/bin/minimal.rs
@@ -43,7 +43,11 @@ mod app {
             errata::clear_int_i2c_interrupt_line,
             timer::{GameTickDriver, Started as TickStarted},
         },
-        game::{tile::LoopingProducer, GameDriver, Message, MAILBOX_CAPACITY},
+        game::{
+            driver::{GameDriver, MAILBOX_CAPACITY},
+            tile::LoopingProducer,
+            Message,
+        },
     };
     use microtile_engine::{gameplay::game::Observer, geometry::grid::Grid};
     use rtic_sync::channel::{Channel, TrySendError};
