+++ src/device/button.rs
@@ -1,4 +1,4 @@
-use crate::game::{Message, MAILBOX_CAPACITY};
+use crate::game::{driver::MAILBOX_CAPACITY, Message};
 use core::marker::PhantomData;
 use microbit::{
     gpio::BTN_B,
